# perf(lexer): Optimize parsing performance with modern JS features